refactor: :truck: Move existing frontend to `apps` folder
