+++ apps/poc-react/.eslintrc.cjs
