+++ apps/poc-react/.prettierrc
