+++ apps/poc-react/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "frontend",
+  "name": "poc-react",
   "private": true,
   "version": "0.0.0",
   "type": "module",
