+++ apps/poc-react/src/components/DiscoverFrame.tsx
