+++ apps/poc-react/src/components/Navbar.tsx
