+++ apps/poc-react/src/components/Notification.tsx
