+++ apps/poc-react/src/components/SelectedTags.tsx
