+++ apps/poc-react/src/components/helper/RequireAuth.tsx
