+++ apps/poc-react/src/components/signUp/SignupInfo.tsx
