+++ apps/poc-react/src/constants.ts
