+++ apps/poc-react/src/vite-env.d.ts
