+++ apps/poc-react/tailwind.config.js
