+++ apps/poc-react/tsconfig.json
