+++ apps/poc-react/tsconfig.node.json
