+++ apps/poc-react/vite.config.ts
